Make FAQ more clear about stopping server
The previous formatting felt a bit confusing. It mentions "just stopping" a server (which implies the mc server) the droplet is still active and will charge you since resources are allocated, just the mc server is off.

Then it mentions:
>When you **stop** a server on Gamocosm, it snapshots (backs up), and destroys your server. 

The "stop" button stops the mc server. The stop button for the actual droplet itself is called "Turn off".

>When you **restart** a server, Gamocosm recreates it from the snapshot.

Similarly, the start button for the droplet is not called start or restart, it's called "Turn on". The restart button exists when the droplet is online, and you want to restart the droplet. Restarting the droplet does not recreate the snapshot from what I know, it just restarts the droplet, it's never deleted.